refactor(repository): refactor repository package
The `Repository` interface has moved into the `playbot` package.

The `MariaDbRepository` implementation has moved into its own `mariadb`
package, and the file was split into many.
